+++ src/prefect/environments/execution/k8s/job.py
@@ -51,17 +51,34 @@ class KubernetesJobEnvironment(Environment):
         on_start: Callable = None,
         on_exit: Callable = None,
     ) -> None:
-        self.identifier_label = str(uuid.uuid4())
         self.job_spec_file = os.path.abspath(job_spec_file) if job_spec_file else None
 
         # Load specs from file if path given, store on object
         self._job_spec = self._load_spec_from_file()
+
+        self._identifier_label = ""
 
         super().__init__(labels=labels, on_start=on_start, on_exit=on_exit)
 
     @property
     def dependencies(self) -> list:
         return ["kubernetes"]
+
+    @property
+    def identifier_label(self) -> str:
+        if not hasattr(self, "_identifier_label") or not self._identifier_label:
+            self._identifier_label = str(uuid.uuid4())
+        return self._identifier_label
+
+    def __getstate__(self) -> dict:
+        state = self.__dict__.copy()
+        # Ensure _identifier_label is not persisted
+        if "_identifier_label" in state:
+            del state["_identifier_label"]
+        return state
+
+    def __setstate__(self, state: dict) -> None:
+        self.__dict__.update(state)
 
     def execute(  # type: ignore
         self, storage: "Docker", flow_location: str, **kwargs: Any