+++ tests/environments/execution/test_k8s_job_environment.py
@@ -80,6 +80,19 @@ def test_create_k8s_job_environment_identifier_label():
         assert environment.identifier_label
 
 
+def test_create_k8s_job_environment_identifier_label_none():
+    with tempfile.TemporaryDirectory() as directory:
+
+        with open(os.path.join(directory, "job.yaml"), "w+") as file:
+            file.write("job")
+
+        environment = KubernetesJobEnvironment(
+            job_spec_file=os.path.join(directory, "job.yaml")
+        )
+        environment._identifier_label = None
+        assert environment.identifier_label
+
+
 def test_setup_k8s_job_environment_passes():
     with tempfile.TemporaryDirectory() as directory:
 
@@ -410,3 +423,9 @@ def test_roundtrip_cloudpickle():
         new = cloudpickle.loads(cloudpickle.dumps(environment))
         assert isinstance(new, KubernetesJobEnvironment)
         assert new._job_spec == "job"
+
+        # Identifer labels do not persist
+        assert environment.identifier_label
+        assert new.identifier_label
+
+        assert environment.identifier_label != new.identifier_label